Simplify LTP/UDP contact volume tests.
Before, the contact volume test procedure was:
For each contactvolume:
 - Start up ION with a config file hand-tuned to provide the contact
   volume.
 - Try to send one bundle.
 - Check if you got it.
 - Close down ION and clean up.

This worked and really clearly dissected the bug.  Now that we
understand and have fixed the bug, we can optimize the test a little
while still maintaining coverage:
 - Start up ION with a config file with no contacts at all
 - Set up the bundle receiver
 - For each contactvolume:
   - add the contact volume
   - send one bundle
   - delete the contact volume
 - Check if you got bundles
 - Close down ION and clean up.

This is faster because ION is really slow to start up or shut down.