Make autotools build compatible with subdir-objs.
This allows the Makefile.am to specify where compiled objects should go,
because an automated test suite won't want to put its compiled objects in
the top-level directory, but the rest of ION will.

This requires expanding all of the $(icibindir), $(iciincludedir) and friends
make variables, to avoid sprinkling dependency tracking in top-level
directories with troublesome names like "$(icibindir)".