Make bp_parse_class_of_service() for common method of specifying C-o-S args.
Before this commit, each ION program could define how it set the
class-of-service arguments in its own way.  bptrace and bpsendfile gathered these
settings from argv in a consistent way, but this way didn't allow setting the
custodySwitch.  They used parseClassOfService() in libbpP, which did bounds
checking but didn't set the extendedCOS structure directly.

This commit moves parseClassOfService() to bp_parse_class_of_service() in
libbp, and extends its syntax to accept custodyRequested and the eCOS flow
label.  It also assigns to the same variables that are passed to bp_send()
to minimize the effort required of the caller.  It also defines a new
BP_PARSE_CLASS_OF_SERVICE_USAGE string that applications can incorporate into
their command-line usage.

Also, bptrace and bpsendfile are updated to use this new helper.