Port 1000.loopback to C, including adding test support libraries.
This commit makes the first ION test in C that is compatible with the new
test suite.  It adds instructions to Makefile.am for future test authors, and
also adds a libtestutil library that has C code that may be helpful for many
tests (like the ionstart() and ionstop() functions).