Fix sdr "may be used uninitialized" warning.
The code path to it being used uninitialized is that the first switch() finds
BP as the UT layer and the second finds a different UT layer.  That won't
happen, but GCC doesn't know that.  So, we initialize SDR to 0 and if we ever
call sdr_begin_xn(), we'll correctly abort since SDR checks whether the sdr reference is 0.