Fix segmentObj "may be used uninitialized" warning.
segmentObj is assigned by insertDataSegment(), which sdr_mallocs it.  If
segmentObj is not allocated, then insertDataSegment() will return either 0
or -1, which will cause handleDataSegment to return before reading the
uninitialized segmentObj.  ios-mac10.5-x86-miles' gcc isn't smart enough to
figure this out, so we initialize segmentObj to 0 (the null SDR pointer) to
suppress the warning.