Fix pduIsFileData "may be used uninitialized" warning.
pduIsFileData is assigned by selectOutPdu(); if selectOutPdu() is not able
to assign pduIsFileData (due to some error), it will return 0, and
cfdpDequeueOutboundPdu() will print an error and return -1 before using
the uninitialized pduIsFileData.  But, gcc on ios-mac10.5-x86-miles isn't
smart enough to figure that out, so we initialize pduIsFileData to suppress
the warning.