Make bping terminate on SIGINT even on slow-to-attach OS X, bping test OS X fix.
On OS X, bpAttach() takes many seconds sometimes.  Before this commit, SIGINT
may be caught by bping before the signal handler is even installed.  This
commit installs the SIGINT handler at the beginning of main() so that bping
terminates quickly on OS X.

Also, the test script assumed that bping will start up and emit a ping in 2
seconds or less, but this is not true on OS X.  This commit fixes that by
increasing the inter-ping interval in the timeout test.