Make "0.0.0.0" token for IP/hostname mean "bind to all interfaces" (INADDR_ANY)
If parseSocketSpec is called with a socketSpec of the form "0.0.0.0[:XXXXX]",
then ipAddress will be set to INADDR_ANY.

Before this commit, a caller of parseSocketSpec would identify failure by
checking if the returned ipAddress is 0.  0 may be INADDR_ANY, a valid s_addr
to pass to bind() when you want to bind to all interfaces, not just one
particular interface.

Now, parseSocketSpec may set ipAddress to 0 upon success, like if the caller's
socketSpec was of the form "0.0.0.0[:XXXXX]".  The caller instead checks the
return code of parseSocketSpec to identify failure.

The tcpcla, stcpcla, udpcla, aoslsa, and udplsa are updated in this commit to
support the new 0.0.0.0 token.  The loopback configurations that use these
adapters are updated to use it.