Issue #260: Teach valgrind about MTAKE/MRELEASE.
MTAKE and MRELEASE allocate/free ION working memory on behalf of ION programs.
They're macros around allocFromIonMemory() and releaseToIonMemory().
To inspect for leaks with valgrind's memcheck tool, the allocator and releaser
are updated to call VALGRIND_{MALLOC,FREE}LIKE_BLOCK() macros, allowing
valgrind to treat MTAKE/MRELEASE like malloc and free to detect leaks.

In order to call VALGRIND_xxx_BLOCK() macros, the HAVE_VALGRIND_VALGRIND_H
macro must be defined; it will be defined by configure on builds that use
autotools (or you can define it yourself if using a plain Makefile).  If
valgrind is not detected, the macros are never called.

A test for this feature is included that verifies that the valgrind macros
detect MTAKE leaks (and don't false-detect MRELEASEd non-leaks).  This test
is skipped if valgrind isn't present.