Add autodoc: automatically append manpages to ION.pdf.
On a system with ghostscript, psutils, groff and groff-base (these are the
Ubuntu package names), the ION.pdf document is automatically generated as
the concatenation of:

1.  doc/ION-beginning.pdf  (currently, pages 1-70 of ION.pdf)
2.  doc/appendix-a-contents.txt (an autogenerated ordered list of the
      manpages and also a version of the repo via "hg id", like:
      480f9503ec1c+ (297-autodoc) ion-2.5.1
3.  doc/ION-manpages.ps (all of the manpages in ps format, via
      man -Tps, organized by section, alphabetically, matching
      appendix-a-contents.txt).

If these tools are not available, ION.pdf is just ION-beginning.pdf :-(.
A warning is printed when this happens.  ./configure will hint the user
as to what packages to install to avoid this.

Your groff must have the "-ms" macro set installed.  This lets the contents
page have multiple columns.  In Ubuntu, this comes in the "groff" package.  I
hear it's installed by default on most other Unixes.  This commit will
auto-detect whether you have "groff -ms" available.