Make autodoc tolerant of autoconf 2.60.  Thanks Josh!
Josh Schendel points out that the minimum autoconf currently required for ION
is 2.60, and so AC_PATH_PROGS_FEATURE_CHECK() may not be available.  This is
the case on some of the buildbots.  This commit replaces
AC_PATH_PROGS_FEATURE_CHECK with a nearly-equivalent shell implementation.  The
replacement implementation cannot check multiple groffs for the -ms macro, but
this feature of _PROGRS_FEATURE_CHECK was unused anyway.