Supply psjoin (small perl script) for psjoin-less systems.
psjoin is used to join together all of the postscripts for turning into
ION.pdf.  Many older systems may not have psjoin.  psjoin is a simple perl
script written by Tom Sato, and released as public domain.

The autoconf mods implement this behavior: If there is a system-wide psjoin,
use that.  If not, use the one supplied here.