The sm_GetUniqueKey() functionality was completely rewritten for Unix process systems that use either SVR4 semaphores or the new Posix Named Semaphores to avoid collisions in the keys.
This Windows Mingw version of the routine is therefore an orphan (and the only known system using the code updated in
this commit).  For that version of the routine, we're using instead the version that Jay Gao proposed as an overall fix.
This should be sufficient to also solve the problems we'd seen on those smaller (single user) platforms.