Address problem from gitlab issue 307
The new Posix Named Semaphores already worked fine if multiple instances of ION were running at the same to on the same computer but ONLY if they were running as the same user.

In theory, multiple users don't matter because the ION semaphores are shared, BUT, the permissions on the semaphores were set to rw-------, so that they couldn't be shared between multip
le users because a NEW user couldn't read or delete the semaphores if the old version of ION crashed (if it exits normally, it cleans itself up).

nts.

It also makes them "slightly" easier to clean up later.

Note that if tighter permissions for a critical ION instance are desired, there are breadcrumbs in the source to set the permissions back they way that they were so that other user proce
sses can't interfere.

(there's also a patch to ./killm to help)This patch changes all of the Posix Named Semaphores to be mode rw-rw-rw-  so that they can truly be shared by multiple processes running concurrently by multiple user accounts.

It also makes them "slightly" easier to clean up later.

Note that if tighter permissions for a critical ION instance are desired, there are breadcrumbs in the source to set the permissions back they way that they were so that other user proce
sses can't interfere.

(there's also a patch to ./killm to help)