Address problem from gitlab issue 307
This patch adds a few lines to (quietly) delete any ION named semaphores on the system.
If they aren't in use, then it does nothing.

It works in Linux and Solaris.  Unfortunately, MacOS doesn't mount them into the file system, so we can't delete THOSE from the script.

If the deletion fails for permission reasons, then the script will list the remaining semaphores and suggest that the user re-run the script
with sudo to clean them up completely.  For some reason, on Linux, only root can delete the semaphores belonging to OTHER users regardless
of their permission bits.