Fixed ionTerminate() segfault and shared memory shutdown
Reordered ionTerminate() so that SDR working memory is detached after SDR
cleanup operations that access the SDR are done. This fixes a segfault when
running the ionexit utility.

Also added marking for destruction the SDR working memory and global semaphore
table shared memory segments during ionTerminate(); and in ionexit, added
stopping of BSSP and CFDP daemons so that the number of processes attached to
shared memory goes to 0 and the memory segments get destroyed after running
ionexit.