Resolved stack-buffer-overflow error by passing `def` instead of `&def` in bslpol_proc_applyReceiverPolRule().
Note: previously, the code passed the address of the local pointer variable(`&def`) instead of the pointer to the `sc_Def` object itself. This caused an out-of-bounds read in `bpsec_sci_stateInit()`. Changing the call to pass `def` correctly resolves the ASan error and improves bpsec test stability while running under ASan.