Changed ionexit_test to verify specifically ION shmem no longer exists
Previously, the test was verifying that the system-wide set of shared memory
was the same before and after the test.

This change makes the verification of ION shared memory segment deallocation
robust against other processes in the system creating or destroying their own
memory segments during the test.

Also realized that the test script can't actually verify that ION does not
destroy other processes' shared memory segments, because it cannot distinguish
whether it was ION or the other process that destroyed the non-ION segment.
Removed that claim from the test script comments.