platform: Check for execinfo.h with AC_CHECK_HEADERS
ION's use of preprocessor statements is intended to only include
<execinfo.h> on Linux when glibc is used. But the check is not
exhaustive (fails for musl) and is fragile (uses historic
system-specific macros that are not in the reserved namespace). Using
Autoconf to check for the header is more intentful and robust.

FreeBSD has had the backtrace() library of functions since 10.0, but
since that implementation differs from glibc's, we'd have to test if
those differences matter in ION's use case, so I've kept
printStackTrace()'s usage of backtrace() reserved for Linux.