nm/contrib: QCBOR: Avoid implicit int-float conversion warning
Compiling with Clang fails because of a statement in qcbor_decode.c
which compares a double to INT64_MAX. This implicitly converts INT64_MAX
from 2^63 - 1 to 2^63 and raises a warning which makes the compilation
fail because of -Werror.

Upstream QCBOR already addressed this back in 2020 (I use their
reasoning to subtract by 0x7ff). Our check here is not comprehensive
(doesn't address NaN or underflow) unlike upstream QCBOR, it's only
enough to make Clang be happy.