Fix bpcp Remote_Remote transfer race condition on exit
In Remote_Remote mode, bpcp was exiting before the last proxy put
request could be processed. The issue was that bpcp waited only for
CfdpEofSentInd (proxy request sent) but not for the proxy request
to be actually received and processed by handleProxyPutRequest().

For the last file in a recursive copy, bpcp would exit immediately
after the proxy request FDU was sent, potentially terminating the
receiver thread before it could call cfdp_get_event() to process
the CfdpMetadataRecvInd event containing the proxy put request.

Changes:
- Add proxy_wait and proxy_done states to wait_status enum (bpcp.h)
- After CfdpEofSentInd, wait up to 2 seconds for CfdpMetadataRecvInd
  to confirm the proxy request was processed (bpcp.c ion_cfdp_rput)
- Signal proxy_done in rcv_msg_thread when matching CfdpMetadataRecvInd
  is received (bpcp.c rcv_msg_thread)
- Send error response on silent createFDU failure in
  handleProxyPutRequest (libcfdpops.c)

This also includes the previously committed fix for the dangling
pointer bug in open_remote_dir() where dirlst.destFileName was
pointing to a stack variable.